Implement SVG export functionality to convert drawing commands into SVG text. (#367)

+++ test/src/CanvasTest.cpp
@@ -626,12 +626,14 @@ TGFX_TEST(CanvasTest, simpleShape) {
   auto point = Point::Make(width / 2, height / 2);
   auto radius = image->width() / 2;
   auto rect = Rect::MakeWH(radius * 2, radius * 2);
-  canvas->drawCircle(point, radius + 30, paint);
-  canvas->setMatrix(Matrix::MakeTrans(point.x - radius, point.y - radius));
+  canvas->drawCircle(point, static_cast<float>(radius) + 30.0f, paint);
+  canvas->setMatrix(Matrix::MakeTrans(point.x - static_cast<float>(radius),
+                                      point.y - static_cast<float>(radius)));
   canvas->drawRoundRect(rect, 10, 10, paint);
 
-  canvas->setMatrix(Matrix::MakeTrans(point.x - radius, point.y - radius));
-  canvas->rotate(45, radius, radius);
+  canvas->setMatrix(Matrix::MakeTrans(point.x - static_cast<float>(radius),
+                                      point.y - static_cast<float>(radius)));
+  canvas->rotate(45.0f, static_cast<float>(radius), static_cast<float>(radius));
   canvas->drawImage(image, SamplingOptions(FilterMode::Linear));
   EXPECT_TRUE(Baseline::Compare(surface, "CanvasTest/shape"));
 }
